Cleaned up and commented FarmLand. Also changed some int variables to long variables since the total weight across all farms could cause overflow if using ints

+++ application/AnnualReport.java
@@ -27,6 +27,7 @@ package application;
 import java.util.ArrayList;
 
 import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -121,12 +122,12 @@ public class AnnualReport extends Stage {
 	 */
 	public static class FarmRow {
 		private SimpleStringProperty farm;
-		private SimpleIntegerProperty weights;
+		private SimpleLongProperty weights;
 		private SimpleStringProperty percentage;
 
-		private FarmRow(String farm, Integer weights, String percentage) {
+		private FarmRow(String farm, Long weights, String percentage) {
 			this.farm = new SimpleStringProperty(farm);
-			this.weights = new SimpleIntegerProperty(weights);
+			this.weights = new SimpleLongProperty(weights);
 			this.percentage = new SimpleStringProperty(percentage);
 
 		}
@@ -139,7 +140,7 @@ public class AnnualReport extends Stage {
 			this.farm.set(farm);
 		}
 
-		public Integer getWeights() {
+		public Long getWeights() {
 			return weights.get();
 		}
 
@@ -168,14 +169,14 @@ public class AnnualReport extends Stage {
 	private static ObservableList<FarmRow> getData(FarmLand farmLand, Integer yearNum) {
 		ArrayList<FarmRow> farmList = new ArrayList<FarmRow>();
 
-		int totalWeight = farmLand.getAllFarmsYearTotal(yearNum);
+		long totalWeight = farmLand.getAllFarmsYearTotal(yearNum);
 
 		for (String farm : farmLand.getFarms()) {
 
 			// Calculate monthly weight and percentage of annual weight for this
 			// month. The Double.MIN_NORMAL is added to the denominator to avoid
 			// divide by zero errors.
-			int weight = farmLand.getFarm(farm).getYearTotal(yearNum);
+			long weight = farmLand.getFarm(farm).getYearTotal(yearNum);
 			double percentage = 100 * weight / (totalWeight + Double.MIN_NORMAL);
 
 			// Convert the month, weight, and percentage to strings for display
